implement creating schedule

+++ src/main/java/com/agnieszka/reservationapp/vetapp/appUser/AppUserRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 @Repository
 @Transactional(readOnly = true)
-public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
+public interface AppUserRepository extends JpaRepository<AppUser, Long> {
 
     Optional<AppUser> findByEmail(String email);
